Develop

+++ src/main/java/com/main/numOps/domain/providers/ProviderRepository.java
@@ -2,6 +2,6 @@ package com.main.numOps.domain.providers;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProviderRepository extends JpaRepository<ProviderModel,Integer> {
+public interface ProviderRepository extends JpaRepository<ProviderModel,Long> {
 
 }
